requirements.txt
huggingface_hub==0.23.4
matplotlib==3.9.1
numpy==2.0.1
pandas==2.2.2
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
